requirements.txt
python-dotenv
streamlit
langchain
langchain-openai
langchain-core
langchain-community
langgraph
faiss-cpu
notebook
Flask
Pillow
requests
numpy
